Tooltips finalized (for now)
Tooltips have largely been reworked, but are much simpler to create and more reliable.

Hovering over something that needs a tooltip shown for it can call UserInterface.mouse_hover to add itself to the hovered_targets array, and reverse for UserInterface.mouse_dehover. The first ([0]) object in the hovered_targets array will take tooltip priority. As long as anything remains in hovered_targets, a tooltip will be shown.

The tooltip itself is responsible for parsing content passed to it before display.

Some potential future improvements:
- Tooltip content is currently redrawn on every frame. Not sure if this is a big deal, but should probably only redraw when the content changes.
- The mouse_hover and mouse_dehover functions currently reference the *entire* object, when only a few relevant strings are needed. This should probably be corrected later.

Overall, much happier with the tooltip system now, so I'm moving on to something else.